DeviceDocNormalizer: split models out and extract pairing helpers
DeviceDocNormalizer.swift was 646 lines mixing six DTO structs with the
parsing/normalization enum. Splits into three focused files:

- DeviceDocumentationModels.swift (119) — the six tightly-coupled DTOs:
  DeviceDocumentation, NormalizedDeviceDoc, DeviceDocIdentity,
  DevicePairingGuide, DevicePairingMethod, DeviceDocCapability. Per the
  Code-structure rule, DTO bundles can stay together.
- DeviceDocNormalizer.swift (361) — the enum entry point, the main
  normalize(parsed:device:) orchestrator, capability/expose/section
  helpers, and the generic span helpers (firstParagraph, plainText,
  matchesAny, etc.) shared across pairing and capability extraction.
- DeviceDocNormalizer+Pairing.swift (175) — pairing-specific helpers
  (extractFromNotes, makePairingGuide, collectStepItems, isPureStepList,
  collectSubsections, defaultPairingSummary) plus the prerequisite /
  success / troubleshooting keyword arrays they use.

A handful of formerly-private helpers (normalizeTitle,
isPairingAdjacentTitle, firstParagraph, collectParagraphSpans, unique,
plainText, matchesAny) bumped to internal so the +Pairing extension
across files can call them. No external API surface change — they were
already module-internal, just file-private within the original file.

The +Pairing path is quoted in pbxproj (the `+` requires it).

Refs #34
