+++ Shellbee/Core/Parsing/DeviceDocNormalizer.swift
@@ -1,119 +1,5 @@
 import Foundation
 
-struct DeviceDocumentation: Sendable {
-    let sourcePath: String
-    let parsed: ParsedDeviceDoc
-    let normalized: NormalizedDeviceDoc
-}
-
-struct NormalizedDeviceDoc: Sendable {
-    let identity: DeviceDocIdentity
-    let pairing: DevicePairingGuide?
-    let capabilities: [DeviceDocCapability]
-    let options: [DocOption]
-    let notesSections: [DocSection]
-    let advancedSections: [DocSection]
-    let miscSections: [DocSection]
-    let quality: Quality
-
-    enum Quality: Sendable, Equatable {
-        case fullyNormalized
-        case partiallyNormalized
-        case parsedOnly
-    }
-
-    var additionalSections: [DocSection] { advancedSections + miscSections }
-    var hasSemanticContent: Bool {
-        pairing != nil || !capabilities.isEmpty || !options.isEmpty || !notesSections.isEmpty
-    }
-}
-
-struct DeviceDocIdentity: Sendable {
-    let vendor: String
-    let model: String
-    let description: String
-    let imageURL: URL?
-    let supportsOTA: Bool
-    let exposesSummary: String?
-}
-
-struct DevicePairingGuide: Sendable {
-    let summary: [InlineSpan]
-    let prerequisites: [[InlineSpan]]
-    let primarySteps: [StepItem]
-    let alternatives: [DevicePairingMethod]
-    let successCues: [[InlineSpan]]
-    let troubleshooting: [[InlineSpan]]
-    let additionalNotes: [DocBlock]
-
-    nonisolated var hasContent: Bool {
-        !summary.isEmpty
-            || !prerequisites.isEmpty
-            || !primarySteps.isEmpty
-            || !alternatives.isEmpty
-            || !successCues.isEmpty
-            || !troubleshooting.isEmpty
-            || !additionalNotes.isEmpty
-    }
-}
-
-struct DevicePairingMethod: Sendable, Identifiable {
-    let id: UUID
-    let title: String
-    let summary: [InlineSpan]
-    let steps: [StepItem]
-    let notes: [DocBlock]
-    /// True when this alternative is purely a reference to the Touchlink guide with no
-    /// device-specific steps. The UI replaces the generic card with an in-app Touchlink button.
-    let isTouchlinkReset: Bool
-    /// True when this alternative describes a Philips Hue serial-number factory reset.
-    /// The UI replaces the raw Z2M content with an in-app Philips Hue Reset action.
-    let isPhilipsHueSerialReset: Bool
-
-    nonisolated init(title: String, summary: [InlineSpan] = [], steps: [StepItem] = [], notes: [DocBlock] = [], isTouchlinkReset: Bool = false, isPhilipsHueSerialReset: Bool = false) {
-        self.id = UUID()
-        self.title = title
-        self.summary = summary
-        self.steps = steps
-        self.notes = notes
-        self.isTouchlinkReset = isTouchlinkReset
-        self.isPhilipsHueSerialReset = isPhilipsHueSerialReset
-    }
-}
-
-struct DeviceDocCapability: Sendable, Identifiable {
-    let id: UUID
-    let title: String
-    let subtitle: String?
-    let summary: String
-    let kind: String
-    let unit: String?
-    let isReadable: Bool
-    let isWritable: Bool
-    let detailChips: [String]
-
-    nonisolated init(
-        title: String,
-        subtitle: String? = nil,
-        summary: String,
-        kind: String,
-        unit: String? = nil,
-        isReadable: Bool,
-        isWritable: Bool,
-        detailChips: [String] = []
-    ) {
-        self.id = UUID()
-        self.title = title
-        self.subtitle = subtitle
-        self.summary = summary
-        self.kind = kind
-        self.unit = unit
-        self.isReadable = isReadable
-        self.isWritable = isWritable
-        self.detailChips = detailChips
-    }
-}
-
 enum DeviceDocNormalizer {
     static nonisolated func normalize(parsed: ParsedDeviceDoc, device: Device) -> NormalizedDeviceDoc {
         // The Z2M device definition is authoritative for connected devices, but catalog
@@ -214,7 +100,7 @@ enum DeviceDocNormalizer {
         )
     }
 
-    private static nonisolated func normalizeTitle(_ title: String) -> String {
+    static nonisolated func normalizeTitle(_ title: String) -> String {
         title
             .trimmingCharacters(in: .whitespacesAndNewlines)
             .lowercased()
@@ -272,158 +158,7 @@ enum DeviceDocNormalizer {
         }
     }
 
-    private static nonisolated func extractFromNotes(_ section: DocSection) -> (pairingBlocks: [DocBlock], pairingRelatedBlocks: [DocBlock], noteBlocks: [DocBlock]) {
-        var pairingBlocks: [DocBlock] = []
-        var pairingRelatedBlocks: [DocBlock] = []
-        var noteBlocks: [DocBlock] = []
-
-        for block in section.blocks {
-            switch block {
-            case .subsection(let title, let blocks):
-                let normalizedTitle = normalizeTitle(title)
-                if normalizedTitle == "pairing" {
-                    pairingBlocks.append(contentsOf: blocks)
-                } else if isPairingAdjacentTitle(normalizedTitle) {
-                    pairingRelatedBlocks.append(.subsection(title: title, blocks: blocks))
-                } else {
-                    noteBlocks.append(.subsection(title: title, blocks: blocks))
-                }
-            default:
-                noteBlocks.append(block)
-            }
-        }
-
-        return (pairingBlocks, pairingRelatedBlocks, noteBlocks)
-    }
-
-    private static nonisolated func makePairingGuide(
-        from pairingBlocks: [DocBlock],
-        relatedBlocks: [DocBlock],
-        identity: DeviceDocIdentity
-    ) -> DevicePairingGuide? {
-        // Only count steps that came from real step-list blocks, not paragraph fallback.
-        // Paragraph fallback produces step text identical to the summary, which would show twice.
-        let primarySteps = collectStepItems(from: pairingBlocks, paragraphFallback: false)
-        let summary = firstParagraph(in: pairingBlocks) ?? defaultPairingSummary(identity: identity, hasSteps: !primarySteps.isEmpty)
-        let summaryText = plainText(summary)
-
-        // Exclude summary so it doesn't also appear in Before You Start / Success / Troubleshooting
-        let paragraphs = collectParagraphSpans(from: pairingBlocks + relatedBlocks)
-            .filter { plainText($0) != summaryText }
-
-        let prerequisites = unique(paragraphs.filter { matchesAny($0, keywords: prerequisiteKeywords) })
-        let successCues = unique(paragraphs.filter { matchesAny($0, keywords: successKeywords) })
-        let troubleshooting = unique(paragraphs.filter { matchesAny($0, keywords: troubleshootingKeywords) })
-
-        // Track every plain-text span already shown in a named section so additionalNotes never repeats them
-        var usedTexts = Set<String>([summaryText])
-        for spans in prerequisites + successCues + troubleshooting {
-            usedTexts.insert(plainText(spans))
-        }
-
-        let alternatives = collectSubsections(from: pairingBlocks)
-            .filter { normalizeTitle($0.title) != "pairing" }
-            .map { subsection in
-                let altSummary = firstParagraph(in: subsection.blocks) ?? []
-                let altSummaryText = plainText(altSummary)
-                let altSteps = collectStepItems(from: subsection.blocks, paragraphFallback: false)
-                let altNotes = subsection.blocks.filter { block in
-                    guard !isPureStepList(block) else { return false }
-                    if case .paragraph(let spans) = block { return plainText(spans) != altSummaryText }
-                    return true
-                }
-                let normalizedSubtitle = normalizeTitle(subsection.title)
-                // A subsection is a pure Touchlink reference when its title mentions Touchlink and it
-                // contains no device-specific steps or notes — just a link to the Touchlink guide.
-                let isTouchlinkReset = normalizedSubtitle.contains("touchlink")
-                    && altSteps.isEmpty
-                    && altNotes.isEmpty
-                // A subsection describes a Philips Hue serial-number reset when it mentions
-                // "touchlink" and "serial". The Z2M docs include raw JSON and frontend references
-                // for this flow which the app replaces with the in-app Philips Hue Reset action.
-                let isPhilipsHueSerialReset = normalizedSubtitle.contains("touchlink")
-                    && normalizedSubtitle.contains("serial")
-                return DevicePairingMethod(
-                    title: subsection.title,
-                    summary: altSummary,
-                    steps: altSteps,
-                    notes: altNotes,
-                    isTouchlinkReset: isTouchlinkReset,
-                    isPhilipsHueSerialReset: isPhilipsHueSerialReset
-                )
-            }
-            .filter { !$0.summary.isEmpty || !$0.steps.isEmpty || !$0.notes.isEmpty || $0.isTouchlinkReset || $0.isPhilipsHueSerialReset }
-
-        // Subsection titles promoted to Alternatives — skip them in additionalNotes
-        let alternativeTitles = Set(alternatives.map { normalizeTitle($0.title) })
-
-        let additionalNotes = (pairingBlocks + relatedBlocks).filter { block in
-            switch block {
-            case .paragraph(let spans), .note(let spans):
-                return !usedTexts.contains(plainText(spans))
-            case .stepList:
-                return false
-            case .subsection(let title, _):
-                return !alternativeTitles.contains(normalizeTitle(title))
-            default:
-                return true
-            }
-        }
-
-        let guide = DevicePairingGuide(
-            summary: summary,
-            prerequisites: prerequisites,
-            primarySteps: primarySteps,
-            alternatives: alternatives,
-            successCues: successCues,
-            troubleshooting: troubleshooting,
-            additionalNotes: additionalNotes
-        )
-
-        return guide.hasContent ? guide : nil
-    }
-
-    private static nonisolated func collectSubsections(from blocks: [DocBlock]) -> [(title: String, blocks: [DocBlock])] {
-        blocks.compactMap { block in
-            if case .subsection(let title, let blocks) = block {
-                return (title, blocks)
-            }
-            return nil
-        }
-    }
-
-    private static nonisolated func isPureStepList(_ block: DocBlock) -> Bool {
-        if case .stepList = block { return true }
-        return false
-    }
-
-    private static nonisolated func collectStepItems(from blocks: [DocBlock], paragraphFallback: Bool = true) -> [StepItem] {
-        var result: [StepItem] = []
-        var autoNumber = 1
-
-        for block in blocks {
-            switch block {
-            case .stepList(let steps):
-                result.append(contentsOf: steps)
-                autoNumber = max(autoNumber, (steps.last?.number ?? 0) + 1)
-            case .paragraph(let spans) where paragraphFallback && result.isEmpty:
-                result.append(StepItem(number: autoNumber, spans: spans))
-                autoNumber += 1
-            case .subsection(_, let subblocks) where result.isEmpty:
-                let nested = collectStepItems(from: subblocks, paragraphFallback: paragraphFallback)
-                if !nested.isEmpty {
-                    result.append(contentsOf: nested)
-                    autoNumber = max(autoNumber, (nested.last?.number ?? 0) + 1)
-                }
-            default:
-                break
-            }
-        }
-
-        return result
-    }
-
-    private static nonisolated func firstParagraph(in blocks: [DocBlock]) -> [InlineSpan]? {
+    static nonisolated func firstParagraph(in blocks: [DocBlock]) -> [InlineSpan]? {
         for block in blocks {
             switch block {
             case .paragraph(let spans):
@@ -439,7 +174,7 @@ enum DeviceDocNormalizer {
         return nil
     }
 
-    private static nonisolated func collectParagraphSpans(from blocks: [DocBlock]) -> [[InlineSpan]] {
+    static nonisolated func collectParagraphSpans(from blocks: [DocBlock]) -> [[InlineSpan]] {
         var result: [[InlineSpan]] = []
         for block in blocks {
             switch block {
@@ -456,7 +191,7 @@ enum DeviceDocNormalizer {
         return result
     }
 
-    private static nonisolated func unique(_ items: [[InlineSpan]]) -> [[InlineSpan]] {
+    static nonisolated func unique(_ items: [[InlineSpan]]) -> [[InlineSpan]] {
         var seen = Set<String>()
         return items.filter { spans in
             let key = plainText(spans)
@@ -465,7 +200,7 @@ enum DeviceDocNormalizer {
         }
     }
 
-    private static nonisolated func plainText(_ spans: [InlineSpan]) -> String {
+    static nonisolated func plainText(_ spans: [InlineSpan]) -> String {
         spans.map { span in
             switch span {
             case .text(let text), .bold(let text), .italic(let text), .boldItalic(let text), .code(let text):
@@ -479,17 +214,9 @@ enum DeviceDocNormalizer {
         .lowercased()
     }
 
-    private static nonisolated func matchesAny(_ spans: [InlineSpan], keywords: [String]) -> Bool {
+    static nonisolated func matchesAny(_ spans: [InlineSpan], keywords: [String]) -> Bool {
         let text = plainText(spans)
         return keywords.contains { text.contains($0) }
-    }
-
-    private static nonisolated func defaultPairingSummary(identity: DeviceDocIdentity, hasSteps: Bool) -> [InlineSpan] {
-        if hasSteps {
-            return [.text("Follow the steps below to pair the \(identity.model) with Zigbee2MQTT.")]
-        } else {
-            return [.text("Pairing guidance for the \(identity.model) is limited. Review the notes below for device-specific instructions.")]
-        }
     }
 
     private static nonisolated func makeCapabilities(from exposes: [Expose]) -> [DeviceDocCapability] {
@@ -621,7 +348,7 @@ enum DeviceDocNormalizer {
             || title.contains("warning")
     }
 
-    private static nonisolated func isPairingAdjacentTitle(_ title: String) -> Bool {
+    static nonisolated func isPairingAdjacentTitle(_ title: String) -> Bool {
         title.contains("troubleshooting")
             || title.contains("factory reset")
             || title.contains("install code")
@@ -631,16 +358,4 @@ enum DeviceDocNormalizer {
             || title.contains("pair")
     }
 
-    private static nonisolated let prerequisiteKeywords = [
-        "coordinator", "adapter", "wake", "awake", "close to", "battery", "install code",
-        "permit join", "bridge", "factory reset"
-    ]
-    private static nonisolated let successKeywords = [
-        "when connected", "turns off", "flash", "flashes", "blink", "blinks", "pulsate",
-        "joined", "success", "light turns", "beep"
-    ]
-    private static nonisolated let troubleshootingKeywords = [
-        "troubleshooting", "issue", "doesn't", "didn't", "retry", "remove the device",
-        "join it again", "re-pair", "not work", "work around"
-    ]
 }
